refactor(voyager): clean up transaction batch plugin
