+++ voyager/plugins/transaction-batch/src/lib.rs
@@ -9,41 +9,27 @@ use std::{
 use either::Either;
 use futures::{stream::FuturesOrdered, StreamExt};
 use ibc_classic_spec::IbcClassic;
-use ibc_union_spec::{Channel, ChannelState, IbcUnion};
+use ibc_union_spec::IbcUnion;
 use itertools::Itertools;
 use jsonrpsee::{
     core::{async_trait, RpcResult},
-    types::ErrorObject,
     Extensions,
 };
 use serde::{Deserialize, Serialize};
 use tracing::{debug, error, info, instrument, trace, warn};
-use unionlabs::{
-    ibc::core::{
-        client::height::Height,
-        commitment::merkle_prefix::MerklePrefix,
-        connection::{
-            self, connection_end::ConnectionEnd, msg_connection_open_try::MsgConnectionOpenTry,
-        },
-    },
-    id::{ClientId, ConnectionId},
-    primitives::Bytes,
-    traits::Member,
-    ErrorReporter, DELAY_PERIOD,
-};
+use unionlabs::{ibc::core::client::height::Height, id::ClientId, traits::Member, ErrorReporter};
 use voyager_message::{
     call::WaitForHeight,
-    data::{ChainEvent, Data, IbcDatagram},
+    data::{ChainEvent, Data},
     filter::simple_take_filter,
     module::{PluginInfo, PluginServer},
     primitives::{ChainId, IbcSpec, QueryHeight},
     DefaultCmd, ExtensionsExt, Plugin, PluginMessage, RawClientId, VoyagerClient, VoyagerMessage,
-    FATAL_JSONRPC_ERROR_CODE,
 };
 use voyager_vm::{call, data, pass::PassResult, seq, BoxDynError, Op};
 
 use crate::{
-    call::{MakeMsg, MakeTransactionBatchesWithUpdate, ModuleCall},
+    call::{MakeTransactionBatchesWithUpdate, ModuleCall},
     callback::ModuleCallback,
     data::{BatchableEvent, EventBatch, EventClassic, EventUnion, ModuleData},
 };
@@ -345,11 +331,8 @@ impl PluginServer<ModuleCall, ModuleCallback> for Module {
             ModuleCall::MakeTransactionBatchesWithUpdateUnion(mk) => {
                 mk.call(self, e.try_get()?).await
             }
-            ModuleCall::MakeMsgV1(make_msg_v1) => do_make_msg_v1(voyager_client, make_msg_v1).await,
-            ModuleCall::MakeMsgUnion(make_msg_union) => {
-                do_make_msg_union(voyager_client, make_msg_union).await
-            } // ModuleCall::WaitForClientUpdateV1(wait) => wait.call(voyager_client).await,
-              // ModuleCall::WaitForClientUpdateUnion(wait) => wait.call(voyager_client).await,
+            ModuleCall::MakeMsgClassic(mk) => mk.call(voyager_client).await,
+            ModuleCall::MakeMsgUnion(mk) => mk.call(voyager_client).await,
         }
     }
 
@@ -377,755 +360,6 @@ impl PluginServer<ModuleCall, ModuleCallback> for Module {
     }
 }
 
-#[instrument(
-    skip_all,
-    fields(
-        %origin_chain_id,
-        %origin_chain_proof_height,
-        %target_chain_id,
-        msg = IbcUnion::event_name(&event)
-    )
-)]
-async fn do_make_msg_union(
-    voyager_client: &VoyagerClient,
-    MakeMsg {
-        origin_chain_id,
-        origin_chain_proof_height,
-        target_chain_id,
-        event,
-    }: MakeMsg<IbcUnion>,
-) -> RpcResult<Op<VoyagerMessage>> {
-    match event {
-        EventUnion::ConnectionOpenInit(connection_open_init_event) => {
-            let client_id = connection_open_init_event.client_id;
-            let counterparty_client_id = connection_open_init_event.counterparty_client_id;
-            let connection_id = connection_open_init_event.connection_id;
-
-            // info of the client on the target chain that will verify the storage
-            // proofs
-            let target_client_info = voyager_client
-                // counterparty_client_id from open_init/try is the client on the target chain
-                .client_info::<IbcUnion>(target_chain_id.clone(), counterparty_client_id)
-                .await?;
-
-            debug!(
-                %counterparty_client_id,
-                %target_client_info.client_type,
-                %target_client_info.ibc_interface,
-                %target_client_info.metadata,
-            );
-
-            // info of the client on the origin chain, this is used to decode the stored
-            // client state
-            let origin_client_info = voyager_client
-                // client_id from open_init/try is the client on the origin chain
-                .client_info::<IbcUnion>(origin_chain_id.clone(), client_id)
-                .await?;
-
-            debug!(
-                %client_id,
-                %origin_client_info.client_type,
-                %origin_client_info.ibc_interface,
-                %origin_client_info.metadata,
-            );
-
-            // the connection end as stored by the origin chain after open_init/try
-            let connection_state = voyager_client
-                .maybe_query_ibc_state(
-                    origin_chain_id.clone(),
-                    origin_chain_proof_height.into(),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .state
-                .ok_or(ErrorObject::owned(
-                    FATAL_JSONRPC_ERROR_CODE,
-                    "connection must exist",
-                    None::<()>,
-                ))?;
-            debug!(
-                connection_state = %serde_json::to_string(&connection_state).unwrap(),
-            );
-
-            // proof of connection_state, encoded for the client on the target chain
-            let connection_proof = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id.clone(),
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .proof;
-            debug!(%connection_proof);
-
-            let encoded_connection_state_proof = voyager_client
-                .encode_proof::<IbcUnion>(
-                    target_client_info.client_type.clone(),
-                    target_client_info.ibc_interface.clone(),
-                    connection_proof,
-                )
-                .await?;
-            debug!(%encoded_connection_state_proof);
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgConnectionOpenTry {
-                        client_id: connection_open_init_event.counterparty_client_id,
-                        counterparty_client_id: connection_open_init_event.client_id,
-                        counterparty_connection_id: connection_open_init_event.connection_id,
-                        proof_height: origin_chain_proof_height.height(),
-                        proof_init: encoded_connection_state_proof,
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::ConnectionOpenTry(connection_open_try_event) => {
-            let client_id = connection_open_try_event.client_id;
-            let counterparty_client_id = connection_open_try_event.counterparty_client_id;
-            let connection_id = connection_open_try_event.connection_id;
-
-            // info of the client on the target chain that will verify the storage
-            // proofs
-            let target_client_info = voyager_client
-                // counterparty_client_id from open_init/try is the client on the target chain
-                .client_info::<IbcUnion>(target_chain_id.clone(), counterparty_client_id)
-                .await?;
-
-            debug!(
-                %counterparty_client_id,
-                %target_client_info.client_type,
-                %target_client_info.ibc_interface,
-                %target_client_info.metadata,
-            );
-
-            // info of the client on the origin chain, this is used to decode the stored
-            // client state
-            let origin_client_info = voyager_client
-                // client_id from open_init/try is the client on the origin chain
-                .client_info::<IbcUnion>(origin_chain_id.clone(), client_id)
-                .await?;
-
-            debug!(
-                %client_id,
-                %origin_client_info.client_type,
-                %origin_client_info.ibc_interface,
-                %origin_client_info.metadata,
-            );
-
-            // the connection end as stored by the origin chain after open_init/try
-            let connection_state = voyager_client
-                .maybe_query_ibc_state(
-                    origin_chain_id.clone(),
-                    origin_chain_proof_height.into(),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .state
-                .ok_or(ErrorObject::owned(
-                    FATAL_JSONRPC_ERROR_CODE,
-                    "connection must exist",
-                    None::<()>,
-                ))?;
-            debug!(
-                connection_state = %serde_json::to_string(&connection_state).unwrap(),
-            );
-
-            // proof of connection_state, encoded for the client on the target chain
-            let connection_proof = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id.clone(),
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .proof;
-            debug!(%connection_proof);
-
-            let encoded_connection_state_proof = voyager_client
-                .encode_proof::<IbcUnion>(
-                    target_client_info.client_type.clone(),
-                    target_client_info.ibc_interface.clone(),
-                    connection_proof,
-                )
-                .await?;
-            debug!(%encoded_connection_state_proof);
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgConnectionOpenAck {
-                        connection_id: connection_open_try_event.counterparty_connection_id,
-                        counterparty_connection_id: connection_open_try_event.connection_id,
-                        proof_height: origin_chain_proof_height.height(),
-                        proof_try: encoded_connection_state_proof,
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::ConnectionOpenAck(connection_open_ack_event) => {
-            let client_id = connection_open_ack_event.client_id;
-            let counterparty_client_id = connection_open_ack_event.counterparty_client_id;
-            let connection_id = connection_open_ack_event.connection_id;
-
-            // info of the client on the target chain that will verify the storage
-            // proofs
-            let target_client_info = voyager_client
-                // counterparty_client_id from open_init/ack is the client on the target chain
-                .client_info::<IbcUnion>(target_chain_id.clone(), counterparty_client_id)
-                .await?;
-
-            debug!(
-                %counterparty_client_id,
-                %target_client_info.client_type,
-                %target_client_info.ibc_interface,
-                %target_client_info.metadata,
-            );
-
-            // info of the client on the origin chain, this is used to decode the stored
-            // client state
-            let origin_client_info = voyager_client
-                // client_id from open_init/ack is the client on the origin chain
-                .client_info::<IbcUnion>(origin_chain_id.clone(), client_id)
-                .await?;
-
-            debug!(
-                %client_id,
-                %origin_client_info.client_type,
-                %origin_client_info.ibc_interface,
-                %origin_client_info.metadata,
-            );
-
-            // the connection end as stored by the origin chain after open_init/ack
-            let connection_state = voyager_client
-                .maybe_query_ibc_state(
-                    origin_chain_id.clone(),
-                    origin_chain_proof_height.into(),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .state
-                .ok_or(ErrorObject::owned(
-                    FATAL_JSONRPC_ERROR_CODE,
-                    "connection must exist",
-                    None::<()>,
-                ))?;
-            debug!(
-                connection_state = %serde_json::to_string(&connection_state).unwrap(),
-            );
-
-            // proof of connection_state, encoded for the client on the target chain
-            let connection_proof = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id.clone(),
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ConnectionPath { connection_id },
-                )
-                .await?
-                .proof;
-            debug!(%connection_proof);
-
-            let encoded_connection_state_proof = voyager_client
-                .encode_proof::<IbcUnion>(
-                    target_client_info.client_type.clone(),
-                    target_client_info.ibc_interface.clone(),
-                    connection_proof,
-                )
-                .await?;
-            debug!(%encoded_connection_state_proof);
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgConnectionOpenConfirm {
-                        connection_id: connection_open_ack_event.counterparty_connection_id,
-                        proof_height: origin_chain_proof_height.height(),
-                        proof_ack: encoded_connection_state_proof,
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::ChannelOpenInit(event) => {
-            let proof_init = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id,
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ChannelPath {
-                        channel_id: event.channel_id,
-                    },
-                )
-                .await?;
-
-            let client_info = voyager_client
-                .client_info::<IbcUnion>(target_chain_id, event.connection.counterparty_client_id)
-                .await?;
-
-            let encoded_proof_init = voyager_client
-                .encode_proof::<IbcUnion>(
-                    client_info.client_type,
-                    client_info.ibc_interface,
-                    proof_init.proof,
-                )
-                .await?;
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgChannelOpenTry {
-                        port_id: event.counterparty_port_id,
-                        channel: Channel {
-                            state: ChannelState::TryOpen,
-                            counterparty_channel_id: Some(event.channel_id),
-                            counterparty_port_id: event.port_id,
-                            connection_id: event.connection.counterparty_connection_id.unwrap(),
-                            version: event.version.clone(),
-                        },
-                        counterparty_version: event.version,
-                        proof_init: encoded_proof_init,
-                        proof_height: origin_chain_proof_height.height(),
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::ChannelOpenTry(event) => {
-            let proof_try = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id,
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ChannelPath {
-                        channel_id: event.channel_id,
-                    },
-                )
-                .await?;
-
-            let client_info = voyager_client
-                .client_info::<IbcUnion>(target_chain_id, event.connection.counterparty_client_id)
-                .await?;
-
-            let encoded_proof_try = voyager_client
-                .encode_proof::<IbcUnion>(
-                    client_info.client_type,
-                    client_info.ibc_interface,
-                    proof_try.proof,
-                )
-                .await?;
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgChannelOpenAck {
-                        channel_id: event.counterparty_channel_id,
-                        counterparty_channel_id: event.channel_id,
-                        counterparty_version: event.version,
-                        proof_try: encoded_proof_try,
-                        proof_height: origin_chain_proof_height.height(),
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::ChannelOpenAck(event) => {
-            let proof_try = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id,
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::ChannelPath {
-                        channel_id: event.channel_id,
-                    },
-                )
-                .await?;
-
-            let client_info = voyager_client
-                .client_info::<IbcUnion>(target_chain_id, event.connection.counterparty_client_id)
-                .await?;
-
-            let encoded_proof_ack = voyager_client
-                .encode_proof::<IbcUnion>(
-                    client_info.client_type,
-                    client_info.ibc_interface,
-                    proof_try.proof,
-                )
-                .await?;
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgChannelOpenConfirm {
-                        channel_id: event.counterparty_channel_id,
-                        proof_ack: encoded_proof_ack,
-                        proof_height: origin_chain_proof_height.height(),
-                    },
-                ),
-            )))
-        }
-
-        EventUnion::PacketSend(event) => {
-            let packet = event.packet();
-
-            let proof_try = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id,
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::BatchPacketsPath::from_packets(&[packet.clone()]),
-                )
-                .await?;
-
-            let client_info = voyager_client
-                .client_info::<IbcUnion>(
-                    target_chain_id,
-                    event.packet.destination_channel.connection.client_id,
-                )
-                .await?;
-
-            let encoded_proof_commitment = voyager_client
-                .encode_proof::<IbcUnion>(
-                    client_info.client_type,
-                    client_info.ibc_interface,
-                    proof_try.proof,
-                )
-                .await?;
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(ibc_union_spec::datagram::MsgPacketRecv {
-                    packets: vec![packet],
-                    relayer_msgs: vec![vec![].into()],
-                    proof: encoded_proof_commitment,
-                    proof_height: origin_chain_proof_height.height(),
-                }),
-            )))
-        }
-
-        EventUnion::WriteAck(event) => {
-            let packet = event.packet();
-
-            let proof_try = voyager_client
-                .query_ibc_proof(
-                    origin_chain_id,
-                    QueryHeight::Specific(origin_chain_proof_height),
-                    ibc_union_spec::path::BatchReceiptsPath::from_packets(&[packet.clone()]),
-                )
-                .await?;
-
-            let client_info = voyager_client
-                .client_info::<IbcUnion>(
-                    target_chain_id,
-                    event.packet.source_channel.connection.client_id,
-                )
-                .await?;
-
-            let encoded_proof_commitment = voyager_client
-                .encode_proof::<IbcUnion>(
-                    client_info.client_type,
-                    client_info.ibc_interface,
-                    proof_try.proof,
-                )
-                .await?;
-
-            Ok(data(IbcDatagram::new::<IbcUnion>(
-                ibc_union_spec::datagram::Datagram::from(
-                    ibc_union_spec::datagram::MsgPacketAcknowledgement {
-                        packets: vec![packet],
-                        acknowledgements: vec![event.acknowledgement],
-                        proof: encoded_proof_commitment,
-                        proof_height: origin_chain_proof_height.height(),
-                    },
-                ),
-            )))
-        }
-    }
-}
-
-async fn do_make_msg_v1(
-    voyager_client: &VoyagerClient,
-    MakeMsg {
-        origin_chain_id,
-        origin_chain_proof_height,
-        target_chain_id,
-        event,
-    }: MakeMsg<IbcClassic>,
-) -> RpcResult<Op<VoyagerMessage>> {
-    match event {
-        EventClassic::ConnectionOpenInit(connection_open_init_event) => {
-            let ConnectionHandshakeStateAndProof {
-                connection_state,
-                encoded_connection_state_proof,
-            } = mk_connection_handshake_state_and_proofs(
-                voyager_client,
-                origin_chain_id,
-                target_chain_id,
-                connection_open_init_event.client_id.clone(),
-                connection_open_init_event.counterparty_client_id.clone(),
-                connection_open_init_event.connection_id.clone(),
-                origin_chain_proof_height,
-            )
-            .await?;
-
-            Ok(data(IbcDatagram::new::<IbcClassic>(
-                ibc_classic_spec::Datagram::from(MsgConnectionOpenTry {
-                    client_id: connection_open_init_event.counterparty_client_id,
-                    counterparty: connection::counterparty::Counterparty {
-                        client_id: connection_open_init_event.client_id,
-                        connection_id: Some(connection_open_init_event.connection_id),
-                        prefix: MerklePrefix {
-                            // TODO: Make configurable
-                            key_prefix: b"ibc".into(),
-                        },
-                    },
-                    // TODO: Make configurable
-                    delay_period: DELAY_PERIOD,
-                    counterparty_versions: connection_state.versions,
-                    proof_height: origin_chain_proof_height,
-                    proof_init: encoded_connection_state_proof,
-                }),
-            )))
-        }
-
-        // MakeMsgV1::MakeMsgConnectionOpenAck(MakeMsgConnectionOpenAck {
-        //     origin_chain_id,
-        //     origin_chain_proof_height,
-        //     target_chain_id,
-        //     connection_open_try_event,
-        // }) => {
-        //     let ConnectionHandshakeStateAndProofs {
-        //         connection_state,
-        //         encoded_connection_state_proof,
-        //         consensus_height,
-        //     } = mk_connection_handshake_state_and_proofs(
-        //         &voyager_client,
-        //         origin_chain_id,
-        //         target_chain_id,
-        //         connection_open_try_event.client_id,
-        //         connection_open_try_event.counterparty_client_id,
-        //         connection_open_try_event.connection_id.clone(),
-        //         origin_chain_proof_height,
-        //     )
-        //     .await?;
-
-        //     Ok(voyager_vm::data(IbcMessage::from(MsgConnectionOpenAck {
-        //         connection_id: connection_open_try_event.counterparty_connection_id,
-        //         counterparty_connection_id: connection_open_try_event.connection_id,
-        //         client_state: encoded_client_state,
-        //         version: connection_state.versions[0].clone(),
-        //         proof_height: origin_chain_proof_height,
-        //         proof_try: encoded_connection_state_proof,
-        //         proof_client: encoded_client_state_proof,
-        //         proof_consensus: encoded_consensus_state_proof,
-        //         consensus_height,
-        //     })))
-        // }
-
-        // MakeMsgV1::MakeMsgConnectionOpenConfirm(MakeMsgConnectionOpenConfirm {
-        //     origin_chain_id,
-        //     origin_chain_proof_height,
-        //     target_chain_id,
-        //     connection_open_ack_event,
-        // }) => {
-        //     // info of the client on the target chain that will verify the storage
-        //     // proofs
-        //     let target_client_info = &voyager_client
-        //         .rpc_server
-        //         // counterparty_client_id from open_try is the client on the target chain
-        //         .client_info(
-        //             &target_chain_id,
-        //             connection_open_ack_event.counterparty_client_id.clone(),
-        //         )
-        //         .await
-        //         ?;
-
-        //     // proof of connection_state, encoded for the client on the target chain
-        //     // this is encoded via the client module for the client on the origin chain
-        //     // (the chain the event was emitted on)
-        //     let connection_proof = &voyager_client
-        //         .rpc_server
-        //         .encode_proof(
-        //             &target_client_info.client_type,
-        //             &target_client_info.ibc_interface,
-        //             &voyager_client
-        //                 .rpc_server
-        //                 .query_ibc_proof(
-        //                     &origin_chain_id,
-        //                     origin_chain_proof_height,
-        //                     ConnectionPath {
-        //                         connection_id: connection_open_ack_event.connection_id.clone(),
-        //                     }
-        //                     .into(),
-        //                 )
-        //                 .await
-        //                 ?
-        //                 .proof,
-        //         )
-        //         .await
-        //         ?;
-
-        //     Ok(voyager_vm::data(IbcMessage::from(
-        //         MsgConnectionOpenConfirm {
-        //             connection_id: connection_open_ack_event.counterparty_connection_id,
-        //             proof_height: origin_chain_proof_height,
-        //             proof_ack: connection_proof,
-        //         },
-        //     )))
-        // }
-
-        // MakeMsgV1::MakeMsgChannelOpenTry(MakeMsgChannelOpenTry {
-        //     origin_chain_id,
-        //     origin_chain_proof_height,
-        //     target_chain_id,
-        //     channel_open_init_event: event,
-        // }) => {
-        //     let origin_channel = voyager_client
-        //         .query_channel(
-        //             origin_chain_id.clone(),
-        //             QueryHeight::Specific(origin_chain_proof_height),
-        //             event.port_id.clone(),
-        //             event.channel_id.clone(),
-        //         )
-        //         .await
-        //         ?;
-
-        //     let proof_init = voyager_client
-        //         .query_ibc_proof(
-        //             &origin_chain_id,
-        //             origin_chain_proof_height,
-        //             ChannelEndPath {
-        //                 port_id: event.port_id.clone(),
-        //                 channel_id: event.channel_id.clone(),
-        //             }
-        //             .into(),
-        //         )
-        //         .await
-        //         ?;
-
-        //     let client_info = voyager_client
-        //         .client_info(&target_chain_id, event.connection.counterparty.client_id)
-        //         .await
-        //         ?;
-
-        //     let encoded_proof_init = voyager_client
-        //         .encode_proof(
-        //             &client_info.client_type,
-        //             &client_info.ibc_interface,
-        //             proof_init.proof,
-        //         )
-        //         .await
-        //         ?;
-
-        //     Ok(data(IbcMessage::from(MsgChannelOpenTry {
-        //         port_id: event.counterparty_port_id,
-        //         channel: Channel {
-        //             state: channel::state::State::Tryopen,
-        //             ordering: origin_channel
-        //                 .state
-        //                 .ok_or(QueueError::Fatal("channel must exist".into()))?
-        //                 .ordering,
-        //             counterparty: channel::counterparty::Counterparty {
-        //                 port_id: event.port_id,
-        //                 channel_id: Some(event.channel_id),
-        //             },
-        //             connection_hops: vec![event.connection.counterparty.connection_id.unwrap()],
-        //             version: event.version.clone(),
-        //             upgrade_sequence: 0,
-        //         },
-        //         counterparty_version: event.version,
-        //         proof_init: encoded_proof_init,
-        //         proof_height: origin_chain_proof_height,
-        //     })))
-        // }
-
-        // MakeMsgV1::MakeMsgChannelOpenAck(MakeMsgChannelOpenAck {
-        //     origin_chain_id,
-        //     origin_chain_proof_height,
-        //     target_chain_id,
-        //     channel_open_try_event,
-        // }) => {
-        //     let origin_channel_path = ChannelEndPath {
-        //         port_id: channel_open_try_event.port_id.clone(),
-        //         channel_id: channel_open_try_event.channel_id.clone(),
-        //     };
-
-        //     let proof_try = voyager_client
-        //         .query_ibc_proof(
-        //             &origin_chain_id,
-        //             origin_chain_proof_height,
-        //             origin_channel_path.into(),
-        //         )
-        //         .await
-        //         ?;
-
-        //     let client_info = voyager_client
-        //         .client_info(
-        //             &target_chain_id,
-        //             channel_open_try_event.connection.counterparty.client_id,
-        //         )
-        //         .await
-        //         ?;
-
-        //     let encoded_proof_try = voyager_client
-        //         .encode_proof(
-        //             &client_info.client_type,
-        //             &client_info.ibc_interface,
-        //             proof_try.proof,
-        //         )
-        //         .await
-        //         ?;
-
-        //     Ok(data(IbcMessage::from(MsgChannelOpenAck {
-        //         port_id: channel_open_try_event.counterparty_port_id,
-        //         channel_id: channel_open_try_event.counterparty_channel_id,
-        //         counterparty_channel_id: channel_open_try_event.channel_id,
-        //         counterparty_version: channel_open_try_event.version,
-        //         proof_try: encoded_proof_try,
-        //         proof_height: origin_chain_proof_height,
-        //     })))
-        // }
-
-        // MakeMsgV1::MakeMsgChannelOpenConfirm(MakeMsgChannelOpenConfirm {
-        //     origin_chain_id,
-        //     origin_chain_proof_height,
-        //     target_chain_id,
-        //     channel_open_ack_event,
-        // }) => {
-        //     let origin_channel_path = ChannelEndPath {
-        //         port_id: channel_open_ack_event.port_id.clone(),
-        //         channel_id: channel_open_ack_event.channel_id.clone(),
-        //     };
-
-        //     let proof_ack = voyager_client
-        //         .query_ibc_proof(
-        //             &origin_chain_id,
-        //             origin_chain_proof_height,
-        //             origin_channel_path.into(),
-        //         )
-        //         .await
-        //         ?;
-
-        //     let client_info = voyager_client
-        //         .client_info(
-        //             &target_chain_id,
-        //             channel_open_ack_event.connection.counterparty.client_id,
-        //         )
-        //         .await
-        //         ?;
-
-        //     let encoded_proof_ack = voyager_client
-        //         .encode_proof(
-        //             &client_info.client_type,
-        //             &client_info.ibc_interface,
-        //             proof_ack.proof,
-        //         )
-        //         .await
-        //         ?;
-
-        //     Ok(voyager_vm::data(IbcMessage::from(MsgChannelOpenConfirm {
-        //         port_id: channel_open_ack_event.counterparty_port_id,
-        //         channel_id: channel_open_ack_event.counterparty_channel_id,
-        //         proof_ack: encoded_proof_ack,
-        //         proof_height: origin_chain_proof_height,
-        //     })))
-        // }
-
-        // MakeMsgV1::MakeMsgRecvPacket(msg) => make_msg_recv_packet(ctx, msg).await,
-        _ => todo!(),
-    }
-}
-
 impl Module {
     #[allow(clippy::type_complexity)] // if you knew why this was here you'd leave me alone
     fn run_pass_internal<'a>(
@@ -1134,7 +368,7 @@ impl Module {
         msgs: Vec<Op<VoyagerMessage>>,
     ) -> Pin<Box<dyn Future<Output = RpcResult<PassResult<VoyagerMessage>>> + Send + 'a>> {
         Box::pin(async move {
-            let mut batchers_v1 =
+            let mut batchers_classic =
                 HashMap::<ClientId, Vec<(usize, BatchableEvent<IbcClassic>)>>::new();
             let mut batchers_union =
                 HashMap::<ibc_union_spec::ClientId, Vec<(usize, BatchableEvent<IbcUnion>)>>::new();
@@ -1166,15 +400,18 @@ impl Module {
 
                             trace!(%client_id, "batching event");
 
-                            batchers_v1.entry(client_id.clone()).or_default().push((
-                                idx,
-                                BatchableEvent {
-                                    first_seen_at,
-                                    provable_height: chain_event.provable_height,
-                                    // TODO: Handle this more gracefully
-                                    event: full_ibc_event.try_into().unwrap(),
-                                },
-                            ));
+                            batchers_classic
+                                .entry(client_id.clone())
+                                .or_default()
+                                .push((
+                                    idx,
+                                    BatchableEvent {
+                                        first_seen_at,
+                                        provable_height: chain_event.provable_height,
+                                        // TODO: Handle this more gracefully
+                                        event: full_ibc_event.try_into().unwrap(),
+                                    },
+                                ));
                         }
 
                         if let Some(full_ibc_event) = chain_event.decode_event::<IbcUnion>() {
@@ -1199,14 +436,14 @@ impl Module {
                     }
                     Err(msg) => {
                         match msg.as_plugin::<ModuleData>(self.plugin_name()) {
-                            Ok(ModuleData::BatchEventsV1(message)) => {
+                            Ok(ModuleData::BatchEventsClassicc(message)) => {
                                 trace!(
                                     client_id = %message.client_id,
                                     events.len = %message.events.len(),
                                     "batching event"
                                 );
 
-                                batchers_v1
+                                batchers_classic
                                     .entry(message.client_id)
                                     .or_default()
                                     .extend(message.events.into_iter().map(|event| (idx, event)));
@@ -1231,7 +468,7 @@ impl Module {
                 };
             }
 
-            let (ready_v1, optimize_further_v1) = batchers_v1
+            let (ready_v1, optimize_further_v1) = batchers_classic
                 .into_iter()
                 .flat_map(|(client_id, events)| split_ready(client_id, events, self))
                 .partition_map::<Vec<_>, Vec<_>, _, _, _>(convert::identity);
@@ -1276,109 +513,6 @@ impl Module {
             })
         })
     }
-}
-
-/// Used to fetch and construct the state and proofs for
-/// MsgConnectionOpenTry/Ack.
-#[instrument(
-    skip_all,
-    fields(
-        %origin_chain_id,
-        %target_chain_id,
-        %client_id,
-        %counterparty_client_id,
-        %connection_id,
-        %origin_chain_proof_height,
-    )
-)]
-async fn mk_connection_handshake_state_and_proofs(
-    voyager_client: &VoyagerClient,
-    origin_chain_id: ChainId,
-    target_chain_id: ChainId,
-    client_id: ClientId,
-    counterparty_client_id: ClientId,
-    connection_id: ConnectionId,
-    origin_chain_proof_height: Height,
-) -> RpcResult<ConnectionHandshakeStateAndProof> {
-    // info of the client on the target chain that will verify the storage
-    // proofs
-    let target_client_info = voyager_client
-        // counterparty_client_id from open_init/try is the client on the target chain
-        .client_info::<IbcClassic>(target_chain_id.clone(), counterparty_client_id.clone())
-        .await?;
-
-    debug!(
-        %counterparty_client_id,
-        %target_client_info.client_type,
-        %target_client_info.ibc_interface,
-        %target_client_info.metadata,
-    );
-
-    // info of the client on the origin chain, this is used to decode the stored
-    // client state
-    let origin_client_info = voyager_client
-        // client_id from open_init/try is the client on the origin chain
-        .client_info::<IbcClassic>(origin_chain_id.clone(), client_id.clone())
-        .await?;
-
-    debug!(
-        %client_id,
-        %origin_client_info.client_type,
-        %origin_client_info.ibc_interface,
-        %origin_client_info.metadata,
-    );
-
-    // the connection end as stored by the origin chain after open_init/try
-    let connection_state = voyager_client
-        .maybe_query_ibc_state(
-            origin_chain_id.clone(),
-            origin_chain_proof_height.into(),
-            ibc_classic_spec::ConnectionPath {
-                connection_id: connection_id.clone(),
-            },
-        )
-        .await?
-        .state
-        .ok_or(ErrorObject::owned(
-            FATAL_JSONRPC_ERROR_CODE,
-            "connection must exist",
-            None::<()>,
-        ))?;
-    debug!(
-        connection_state = %serde_json::to_string(&connection_state).unwrap(),
-    );
-
-    // proof of connection_state, encoded for the client on the target chain
-    let connection_proof = voyager_client
-        .query_ibc_proof(
-            origin_chain_id.clone(),
-            QueryHeight::Specific(origin_chain_proof_height),
-            ibc_classic_spec::ConnectionPath {
-                connection_id: connection_id.clone(),
-            },
-        )
-        .await?
-        .proof;
-    debug!(%connection_proof);
-
-    let encoded_connection_state_proof = voyager_client
-        .encode_proof::<IbcClassic>(
-            target_client_info.client_type.clone(),
-            target_client_info.ibc_interface.clone(),
-            connection_proof,
-        )
-        .await?;
-    debug!(%encoded_connection_state_proof);
-
-    Ok(ConnectionHandshakeStateAndProof {
-        connection_state,
-        encoded_connection_state_proof,
-    })
-}
-
-struct ConnectionHandshakeStateAndProof {
-    connection_state: ConnectionEnd,
-    encoded_connection_state_proof: Bytes,
 }
 
 #[allow(clippy::type_complexity)] // skill issue
